feat: Added Properties() method
Based on a feedback, sometimes verbosely adding all the properties
based on their types can be painful difficult, such as wanting to put
some context object in would result to having to access each properties
individually. The Properties() method is supposed to address that by
doing the parsing internally using the typeof operator.